backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cron_descriptor==2.0.5
cycler==0.12.1
Django==5.2.4
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-decouple==2.1
django-timezone-field==7.1
djangorestframework==3.16.0
et_xmlfile==2.0.0
fake-useragent==2.2.0
firecrawl-py==3.0.3
fonttools==4.59.0
frozenlist==1.7.0
gevent==25.8.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
kiwisolver==1.4.8
kombu==5.5.4
lxml==6.0.0
matplotlib==3.10.3
multidict==6.6.4
mysqlclient==2.2.7
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pandas_ta==0.3.14b0
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
requests==2.32.4
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
squarify==0.4.4
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
vine==5.1.0
vnai==2.1.9
vnstock==3.2.6
vnstock_ezchart==0.0.2
wcwidth==0.2.13
websockets==15.0.1
wordcloud==1.9.4
yarl==1.20.1
zipp==3.23.0
zope.event==5.1.1
zope.interface==7.2
